Greatly improved the speed of the rendering. Changes to portability so as to accept different cameras than the xtionCalib.

+++ src/fancyWindow.h
@@ -48,7 +48,7 @@ private slots:
 
     void on_saveSensorImage_clicked();
 
-    void on_voxelLeaf_valueChanged(int arg1);
+    void on_voxelLeaf_valueChanged(double arg1);
 
     void on_normalRejection_valueChanged(double arg1);
